requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
h3>=3.7.0
flask>=3.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
folium>=0.15.0
gunicorn>=21.0.0
